requirements.txt
pygame
Pillow
bitarray
numpy